requirements.txt
cryptography
qrcode
Pillow
tk
